refactor: DispatcherServlet.service() 리팩토링
