+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/exception/HandlerNotFoundException.java
@@ -1,0 +1,7 @@
+package nextstep.mvc.tobe.exception;
+
+public class HandlerNotFoundException extends RuntimeException {
+    public HandlerNotFoundException() {
+        super();
+    }
+}
